test: fix tests for older node versions.

+++ src/rsa/rsa.c
@@ -1799,7 +1799,7 @@ void
 bcrypto_rsa_key_free(bcrypto_rsa_key_t *key) {}
 
 bcrypto_rsa_key_t *
-bcrypto_rsa_privkey_generate(int bits, int exp) {
+bcrypto_rsa_privkey_generate(int bits, unsigned long long exp) {
   return NULL;
 }
